+++ src/routes/v0/__init__.py
@@ -1,0 +1,5 @@
+from .llm import router as llm
+from .thread import router as thread
+from .tool import router as tool
+
+__all__ = ["llm", "thread", "tool"]
