+++ src/tools/retrieval.py
@@ -6,11 +6,11 @@ from src.utils.retrieval import VectorStore
 from src.constants import DEFAULT_VECTOR_STORE_PATH
 
 @tool
-def retrieval_query(query: str):
+def retrieval_query(query: str, search_type: str = "similarity", search_kwargs: dict = {"k": 10}):
     """Query the vector store. Search type can be 'mmr' or 'similarity'. Search kwargs is a dictionary of kwargs for the search type."""
     loaded_vector_store = VectorStore().load_vector_store()
     vector_store = VectorStore(loaded_vector_store)
-    return vector_store.retrieve(query, "similarity", {"k": 10})
+    return vector_store.retrieve(query, search_type, search_kwargs)
 
 @tool
 def retrieval_add(docs: list[Document]):
